requirements.txt
click==6.6
Flask==0.11.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
numpy==1.11.1
pandas==0.18.1
Pillow==3.3.0
python-dateutil==2.5.3
pytz==2016.6.1
six==1.10.0
uWSGI==2.0.13.1
Werkzeug==0.11.10
